requirements.txt
discord.py>=2.3.0
python-dotenv>=1.0.0
aiofiles>=23.1.0
groq>=0.11.0
aiomysql>=0.2.0
flask>=2.3.0
flask-cors>=4.0.0
asyncio
datetime
typing
re
json
os